# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Social Protection Management Information System API",
        "title": "Guimba-GO API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/beneficiaries": {
            "get": {
                "description": "Returns a paginated list of beneficiaries with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beneficiaries"
                ],
                "summary": "List beneficiaries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status (active, inactive, suspended)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by program ID",
                        "name": "program_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BeneficiaryListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new beneficiary",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beneficiaries"
                ],
                "summary": "Create a beneficiary",
                "parameters": [
                    {
                        "description": "Beneficiary data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateBeneficiaryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.BeneficiaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/beneficiaries/{id}": {
            "get": {
                "description": "Retrieves a beneficiary by ID with program enrollments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beneficiaries"
                ],
                "summary": "Get a beneficiary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Beneficiary ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BeneficiaryResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing beneficiary by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beneficiaries"
                ],
                "summary": "Update a beneficiary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Beneficiary ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated beneficiary data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateBeneficiaryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BeneficiaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Soft-deletes a beneficiary by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beneficiaries"
                ],
                "summary": "Delete a beneficiary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Beneficiary ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/beneficiaries/{id}/programs": {
            "post": {
                "description": "Enrolls a beneficiary in a social program",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beneficiaries"
                ],
                "summary": "Enroll beneficiary in program",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Beneficiary ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Program enrollment",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.EnrollProgramRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BeneficiaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/beneficiaries/{id}/programs/{programId}": {
            "delete": {
                "description": "Removes a beneficiary from a social program",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "beneficiaries"
                ],
                "summary": "Remove beneficiary from program",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Beneficiary ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Program ID (UUID)",
                        "name": "programId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apperror.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperror.Code": {
            "type": "string",
            "enum": [
                "NOT_FOUND",
                "VALIDATION_ERROR",
                "UNAUTHORIZED",
                "FORBIDDEN",
                "CONFLICT",
                "INTERNAL_ERROR",
                "BAD_REQUEST"
            ],
            "x-enum-varnames": [
                "CodeNotFound",
                "CodeValidation",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeConflict",
                "CodeInternal",
                "CodeBadRequest"
            ]
        },
        "apperror.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/apperror.Code"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "apperror.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/apperror.ErrorBody"
                }
            }
        },
...
        "dto.BeneficiaryListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.BeneficiaryResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.BeneficiaryResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "national_id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "programs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProgramEnrollmentResponse"
                    }
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.CreateBeneficiaryRequest": {
            "type": "object",
            "required": [
                "full_name",
                "status"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "national_id": {
                    "type": "string",
                    "maxLength": 100
                },
                "notes": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 50
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "inactive",
                        "suspended"
                    ]
                }
            }
        },
...
        "dto.EnrollProgramRequest": {
            "type": "object",
            "required": [
                "program_id"
            ],
            "properties": {
                "program_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.ProgramEnrollmentResponse": {
            "type": "object",
            "properties": {
                "enrolled_at": {
                    "type": "string"
                },
                "program_id": {
                    "type": "string"
                },
                "program_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateBeneficiaryRequest": {
            "type": "object",
            "required": [
                "full_name",
                "status"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "national_id": {
                    "type": "string",
                    "maxLength": 100
                },
                "notes": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 50
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "inactive",
                        "suspended"
                    ]
                }
            }
        },
...
    }
}